+++ libnd4j/include/ops/declarable/generic/images/adjust_contrast.cpp
@@ -40,8 +40,7 @@ CONFIGURABLE_OP_IMPL(adjust_contrast, 1, 1, true, 0, 0) {
   REQUIRE_TRUE(block.numT() > 0 || block.width() > 1, 0, "ADJUST_CONTRAST: Scale factor required");
   REQUIRE_TRUE(input->rankOf() > 2, 0, "ADJUST_CONTRAST: op expects rank of input array to be >= 3, but got %i instead",
                input->rankOf());
-  //    REQUIRE_TRUE(input->sizeAt(-1) == 3, 0, "ADJUST_CONTRAST: operation expects image with 3 channels (R, G, B), but
-  //    got %i instead", input->sizeAt(-1));
+
 
   NDArray* factor = nullptr;
 
@@ -53,7 +52,7 @@ CONFIGURABLE_OP_IMPL(adjust_contrast, 1, 1, true, 0, 0) {
   }
 
   // fill up axes vector first
-  std::vector<int> axes(input->rankOf() - 1);
+  std::vector<LongType> axes(input->rankOf() - 1);
   for (auto i = 0; i < axes.size(); ++i) axes[i] = i;
 
   // mean as reduction for last dimension set
@@ -99,7 +98,7 @@ CONFIGURABLE_OP_IMPL(adjust_contrast_v2, 1, 1, true, 0, 0) {
     factor->p(0, T_ARG(0));
   }
 
-  std::vector<int> axes({1});  // dim 1 of pseudoresult
+  std::vector<LongType> axes({1});  // dim 1 of pseudoresult
 
   // mean as reduction for last dimension set over size (dim 1) of result3D
   auto mean = input3D.reduceAlongDimension(reduce::Mean, axes);
