+++ libnd4j/include/ops/declarable/helpers/batchnorm.h
@@ -29,7 +29,7 @@ namespace ops {
 namespace helpers {
 
 SD_LIB_HIDDEN void batchnorm(const NDArray* input, const NDArray* mean, const NDArray* variance, const NDArray* gamma,
-                             const NDArray* beta, NDArray* output, const std::vector<int>& axes, const double epsilon);
+                             const NDArray* beta, NDArray* output, const std::vector<LongType>& axes, const double epsilon);
 
 }
 }  // namespace ops
